make the errorprone compiler happy by using an explicit cast to a fp number, because we use a delta equals() method

+++ src/test/java/nl/b3p/tailormap/api/controller/VersionControllerIntegrationTest.java
@@ -91,7 +91,7 @@ class VersionControllerIntegrationTest {
 
         // buildDate can be a few seconds off when running integration tests
         assertEquals(
-                buildDate,
+                (double) buildDate,
                 LocalDateTime.parse(
                                 versionController.getVersion().get("buildDate"),
                                 DateTimeFormatter.ISO_OFFSET_DATE_TIME)
@@ -136,7 +136,7 @@ class VersionControllerIntegrationTest {
                 "Unexpected version response, some keys are missing.");
 
         assertEquals(
-                buildDate,
+                (double) buildDate,
                 LocalDateTime.parse(
                                 versionController.getVersion().get("buildDate"),
                                 DateTimeFormatter.ISO_OFFSET_DATE_TIME)
